+++ tests/apps/core/test_language_chooser.py
@@ -2,6 +2,8 @@
 """
 Unit tests for the language_chooser menu
 """
+from django.test.utils import override_settings
+
 from cms.api import create_page
 from cms.test_utils.testcases import CMSTestCase
 
@@ -11,7 +13,7 @@ from richie.apps.core.helpers import create_i18n_page
 class LanguageChooserTests(CMSTestCase):
     """Test the language chooser menu for available languages"""
 
-    def test_language_chooser_available_language(self):
+    def test_language_chooser_available_language_public(self):
         """
         Menu should always contain every available language from settings,
         no matter if the page has a translation or not
@@ -34,6 +36,45 @@ class LanguageChooserTests(CMSTestCase):
 
         self.assertContains(response, "en, currently in English")
         self.assertContains(response, "fr, switch to French")
+
+    @override_settings(
+        LANGUAGES=(("en", "English"), ("fr", "Français"), ("de", "Deutsch")),
+        CMS_LANGUAGES={
+            1: [
+                {"code": "fr", "hide_untranslated": False, "name": "Français"},
+                {"code": "de", "hide_untranslated": False, "name": "Deutsch"},
+                {
+                    "public": False,
+                    "code": "en",
+                    "hide_untranslated": False,
+                    "name": "English",
+                },
+            ]
+        },
+    )
+    def test_language_chooser_available_language_not_public(self):
+        """
+        Languages that are marked as not public should only be visible to staff users.
+        """
+        content = {
+            "de": "Sprachmenü test",
+            "en": "Language menu test",
+            "fr": "Test du menu de langues",
+        }
+        page = create_i18n_page(
+            content, published=True, template="richie/single_column.html"
+        )
+        url = page.get_absolute_url(language="fr")
+        response = self.client.get(url)
+
+        # English should not be in the language chooser as it is not public
+        self.assertContains(response, page.get_absolute_url(language="fr"))
+        self.assertContains(response, page.get_absolute_url(language="de"))
+        self.assertNotContains(response, page.get_absolute_url(language="en"))
+
+        self.assertContains(response, "fr, actuellement en Français")
+        self.assertContains(response, "de, basculer vers Allemand")
+        self.assertNotContains(response, "en, basculer vers Anglais")
 
     def test_language_chooser_available_language_with_translated_page(self):
         """